+++ NEWS.md
@@ -1,3 +1,4 @@
 # tndata 0.1.0
 
 * Added `get_theme_summaries()` for fetching datasets count per theme.
+* Added `get_datasets()` and `download_dataset()` to enable users to search for available datasets by keyword and download them using their titles.